requirements.txt
python-slugify==4.0.1
hdx-python-api==5.0.3
-r docker-requirements.txt